enterprise-mcp-server/requirements/base.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# MCP SDK
mcp>=1.0.0

# Structured Logging
structlog>=23.2.0

# Metrics
prometheus-client>=0.19.0

# Storage — PostgreSQL
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0

# Storage — Redis
redis>=5.0.0
celery[redis]>=5.3.6

# Rate Limiting
slowapi>=0.1.9

# Retry
tenacity>=8.2.3

# Document Processing
pandas>=2.0.0
openpyxl>=3.1.2
python-docx>=1.1.0
docx2pdf>=0.1.8

# Gmail API
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# HTTP Client
httpx>=0.25.2

# SSE Transport
sse-starlette>=1.6.0

# Environment
python-dotenv>=1.0.0
